[JUNE10-COMMIT-01] added new admin registration api and modified other apis as per the comments form demo

+++ scll-customer-registration/src/main/java/com/scll/customer/registration/CustomerRegistrationFunctionHandler.java
@@ -44,7 +44,7 @@ public class CustomerRegistrationFunctionHandler implements RequestHandler<Custo
         customer.setCustomerMobileNumber(registrationRequest.getMobileNumber());
         customer.setCustomerEmailAddress(registrationRequest.getEmailID());
         customer.setStoreID(registrationRequest.getScllStore());
-        customer.setAvailablePoints(new Double(0));
+        customer.setAvailablePoints(new Double(2000));
         
         
         try {
